microservices/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.1
python-multipart==0.0.6
azure-identity==1.15.0
azure-storage-blob==12.19.0
pyodbc==5.0.1
apscheduler==3.10.4
